refactor: extract file scan and sample helpers into transfer-scan module
